# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for WartegKita — connecting customers with traditional Indonesian Warteg sellers.",
        "title": "WartegKita API",
        "contact": {
            "name": "WartegKita Team",
            "url": "https://github.com/Krisn4novianto/WartegKita_Project"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sellers/{seller_id}/profile": {
            "get": {
                "description": "Returns the internal profile data of a seller (nama warteg, pemilik, bank, dll)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sellers"
                ],
                "summary": "Get Seller Profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Seller ID",
                        "name": "seller_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_krisn4novianto_wartegkita_backend_database_seller.SellerProfile"
                        }
                    }
                }
            },
            "put": {
                "description": "Create or update the seller's internal profile (upsert by seller_id)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sellers"
                ],
                "summary": "Update Seller Profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Seller ID",
                        "name": "seller_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Seller profile payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_krisn4novianto_wartegkita_backend_database_seller.SellerProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_krisn4novianto_wartegkita_backend_database_seller.SellerProfile": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "bank": {
                    "type": "string"
                },
                "deskripsi": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nama_pemilik": {
                    "type": "string"
                },
                "nama_warteg": {
                    "type": "string"
                },
                "nomor_hp": {
                    "type": "string"
                },
                "nomor_rekening": {
                    "type": "string"
                },
                "seller_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token as: Bearer \u003cyour_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}